Branch1-1-part-6
I add another blog and called it (How to prepare for coronavirus: The shopping list for your own home quarantine kit). I add one video and six pictures to the the blog also. "Took a little time learning on how to add the pitcure in a very ok looking way to the site."